% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{basicIndices}
\alias{basicIndices}
\title{updates sample data with basic statistics}
\usage{
basicIndices(x, level = c("VpJ", "V", "J", "VJ", "CDR3aa"))
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{level}{level of reppertoire to assess diversity, choice are VpJ, V, J, VJ or CDR3aa)}
}
\value{
a data frame of Chao
}
\description{
function produces several useful diversity indices.
}
