% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{Chaowor}
\alias{Chaowor}
\title{Adjusted Chao1 for sampling without replacement}
\usage{
Chaowor(x)
}
\arguments{
\item{x}{a vector of counts}
}
\value{
a value of Chao1
}
\description{
function compute the correct
}
