% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{countFeatures}
\alias{countFeatures}
\title{count reads}
\usage{
countFeatures(x, level = c("VpJ", "V", "J", "VJ", "CDR3aa"))
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}

\item{level}{"V", "J" or "VJ" genes.}
}
\value{
a data.table of assay with unique features in rows and samples in columns.
}
\description{
function assay reads by V, J or V-J genes.
}
