% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{divLevel}
\alias{divLevel}
\title{compute diversity indices}
\usage{
divLevel(
  x,
  index = c("shannon", "simpson", "invsimpson"),
  level = c("VJ", "V", "J"),
  norm = TRUE,
  base = exp(1)
)
}
\arguments{
\item{x}{a object of class RepSeqExperiment.}

\item{index}{diversity index, the choice are "shannon", "simpson", "invsimpson".}

\item{level}{diversity index should be computed for the level "V", "J", or "VJ".}

\item{norm}{normalized diversity index will be returned (divided by log(S) with S number of observed species).}

\item{base}{log base . Default value: exp(1)}
}
\value{
a data.table of diversity indices computed at the level "level".
}
\description{
function computes diversity indices for level \code{VJ}, \code{V} or \code{J}
}
