% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{dropSamples}
\alias{dropSamples}
\title{drop sample(s) from a RepSeqExperiment object}
\usage{
dropSamples(x, samples)
}
\arguments{
\item{x}{a RepSeqExperiment object.}

\item{samples}{a vector containing sample names to be removed or a vector of their positions in sampleData slot.}
}
\value{
a RepSeqExperiment object.
}
\description{
function allows to remove one or several samples from an RepSeqExperiment object.
}
