% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{estimateSF}
\alias{estimateSF}
\title{estimate size factor}
\usage{
estimateSF(x, level = c("VpJ", "V", "J", "VJ", "CDR3aa"),
  method = c("GMPR", "Chao", "Chao.gmmean", "Chao.median"))
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{level}{level of repertoire to analyze \code{VpJ, VJ, V, J, CDR3aa}.}

\item{method}{method used for size factor computation}
}
\description{
function
}
