% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filetype}
\alias{filetype}
\title{get type of file}
\usage{
filetype(path)
}
\arguments{
\item{path}{path to a file}
}
\value{
a string indicating file type
}
\description{
function allows to get the type of a file
}
