% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filterassay}
\alias{filterassay}
\title{filter clonotypes based on assay}
\usage{
filterassay(x, n = 1)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}

\item{n}{an integer below this value clonotypes will be filtered out}
}
\value{
an object of class \code{RepSeqExperiment}
}
\description{
function filters out every clonotypes having low assay across all samples.
}
