% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filterClonotypes}
\alias{filterClonotypes}
\title{filter clonotypes}
\usage{
filterClonotypes(
  raw,
  keep.ambiguous = FALSE,
  keep.unproductive = FALSE,
  aa.th = 8,
  outFiltered = F
)
}
\arguments{
\item{raw}{a clonotype table.}

\item{keep.ambiguous}{a boolean choice if ambiguous clonotypes (contain STOP codon) should be kept in analysis. Default is \code{FALSE}.}

\item{keep.unproductive}{a boolean choice if unproductive clonotypes (Euclidean dividion of aa length by 3 > 0) should be kept in analysis. Default is \code{FALSE}.}

\item{aa.th}{an interger indicates the maximal number of amino acids could be deviated from the mean length. Default is \code{8}.}

\item{outFiltered}{write dropped reads to file. Default folder is getwd().}
}
\value{
a filtered clonotype table.
}
\description{
function loads a clonotype table and applies a filter according to several conditions.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
l <- list.files(system.file(file.path('extdata/mixcr'), 
                 package = 'RepSeqData'), 
                 full.names = TRUE)
l
# first element of the vector
path <- l[1]
dataset <- parseMiXCR(path, chain = "B")
dataset
datasetFilter <- filterClonotypes(raw = dataset, 
                                  keep.ambigurous = FALSE, 
                                  keep.ambiguous = FALSE, 
                                  keep.unproductive = FALSE, 
                                  aa.th = 8, 
                                  outFiltered = F)
datasetFilter
}
}
