% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filterFrequency}
\alias{filterFrequency}
\title{filter based on expression.}
\usage{
filterFrequency(x, freq = 0.2, group = NULL)
}
\arguments{
\item{x}{count data set with features in rows and samples in columns.}

\item{freq}{a threshold expressed between 0 & 1 below which data will be filtered out.}

\item{group}{a factor in sampleData, if not null, freq is applied to the group having the smallest number of membres.}
}
\value{
subset of x that match the conditions.
}
\description{
this function allows to filter data according to the percentage of presence across samples.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeq)
library(RepSeqData)
sData(RepSeqData)
# get overlap clonotypes between the sample S01 and the sample S02
filterdata <- getOverlaps(RepSeqData, 
                       freq = 0.2, 
                       group = "quantity")  
filterdata
}
}
