% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getOverlaps}
\alias{getOverlaps}
\title{get shared clonotypes}
\usage{
getOverlaps(x, level = c("VpJ", "CDR3dna"), libnames = NULL)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}

\item{level}{level for shared}
}
\value{
an object of class [\code{\linkS4class{RepSeqExperiment}}]
}
\description{
function returns an RepSeqExperiment object containing shared clonotyes which expressed only in at least two samples
}
