% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getPrivates}
\alias{getPrivates}
\title{get private clonotypes}
\usage{
getPrivates(x)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}
}
\value{
an object of class [\code{\linkS4class{RepSeqExperiment}}]
}
\description{
function returns private clonotyes which expressed only in one sample
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
privateClones <- getPrivates(RepSeqData) # get only private clonotypes
provateClones
}
}
