% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getSingleton}
\alias{getSingleton}
\title{get singletons}
\usage{
getSingleton(x)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}
}
\value{
an object of class \code{RepSeqExperiment}
}
\description{
function gets all clonotype having only 1 count across all samples
}
