% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{muScore}
\alias{muScore}
\title{compute multivariate score}
\usage{
muScore(x, level = c("V", "J", "VJ", "VpJ", "CDR3aa"),
  type = c("count", "usage"))
}
\arguments{
\item{x}{a RepSeqExperiment object}

\item{level}{of repertoire to analyze \code{VJ, V, J, VpJ, CDR3aa}.}

\item{type}{type of data to compute muScore, \code{count} or \code{usage}.}
}
\value{
a data.table ordered by the multivariate decreasing scores. The last column contains the multivariate score.
}
\description{
function computes multivariate score for each feature at level of the repertoire \code{VJ, V, J}. The level VpJ is not supported for the moment.
}
