% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{muScoreM}
\alias{muScoreM}
\title{compute pre-muScore matrix
This function returns a matrix of scores comparing pairwise rows for each sample of the input data set.}
\usage{
muScoreM(df)
}
\arguments{
\item{df}{a data frame of counts with samples in columns and clonotypes in rows.}
}
\value{
a matrix
}
\description{
compute pre-muScore matrix
This function returns a matrix of scores comparing pairwise rows for each sample of the input data set.
}
