% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{muScoreOB}
\alias{muScoreOB}
\title{compute muScore}
\usage{
muScoreOB(df)
}
\arguments{
\item{df}{a data frame of counts with samples in columns and clonotypes in rows}
}
\value{
a vector of multivariate scores computed for each row
}
\description{
This function returns multivariate score for each row of the input mumerical data set using the O'Brien (1984)'s method.
}
