% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{parseClonotypeR}
\alias{parseClonotypeR}
\title{parse ClonotypeR tables}
\usage{
parseClonotypeR(path, chain = c("A", "B"))
}
\arguments{
\item{path}{full path to a TSV file returned by ClonotypeR (tab-delimited text file, could be gzipped).}

\item{chain}{TCR chain \code{A} or [code{B} to extract. Default value is \code{A}.}
}
\value{
a data.table having 7 columns. \code{lib} name of the repertoire, \code{V} V gene identification, \code{J} J gene identification, \code{CDR3aa} CDR3aa chain, \code{CDR3dna} CDR3 DNA chain, \code{score} mapq quality score, \code{count} clonotype assay. Clonotypes were deleted if CDR3aa chain contains STOP codon (*), CDR3dna length is not divisible by 3 or CDR3dna chain contains base "N".
}
\description{
function imports clonotype tables produced by ClonotypeR aligner.
}
\examples{
\dontrun{
dataset <- parseClonotypeR(path, chain="B")
}
}
