% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{parseRTCR}
\alias{parseRTCR}
\title{parse rTCR tables}
\usage{
parseRTCR(path)
}
\arguments{
\item{path}{full path to a file returned by rTCR}
}
\value{
a data.table
}
\description{
function imports clonotype tables produced by rTCR aligner.
}
