% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{perturbation}
\alias{perturbation}
\title{perturbation scores computation}
\usage{
perturbation(
  x,
  ctrl.names,
  distance = c("manhattan", "euclidean", "canberra", "minkowski", "maximum"),
  p = 2
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment.}

\item{ctrl.names}{a vector of characters indicating the names of samples used as control repertoire.}

\item{distance}{distance used for perturbation calculus.}

\item{p}{an integer, the power of Minkowski distance. Default p = 2.}
}
\value{
a data frame containing perturbation scores for each V-genes.
}
\description{
function computes the perturbation scores as a distance between each repertoire and the mean repertoire of the control group.
}
\examples{
\dontrun{
## The package RepSeqData contains example datasets 
library(RepSeqData)
rownames(sData(RepSeqData))
pert <- perturbation(RepSeqData, ctrl.names = c("S01", "S02", "S03"), distance = "manhattan")
pert
}
}
