% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot2v2count}
\alias{plot2v2count}
\title{sample against sample plot}
\usage{
plot2v2count(
  x,
  level = c("V", "J", "VJ", "VpJ", "CDR3aa"),
  libs = NULL,
  scale = c("counts", "log")
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{level}{choose between V, J, VJ, VpJ, CDR3aa}

\item{libs}{a vector of characters of length 2 containing the sample names to plot.}

\item{scale}{normal or log10 scale of axes}
}
\value{
a dot plot
}
\description{
function plots
}
