% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotCountVJ}
\alias{plotCountVJ}
\title{plot a heatmap of V-J count}
\usage{
plotCountVJ(x, sampleName = NULL, scale = c("counts", "percent", "cpm"))
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{sampleName}{sample to plot, sample name must be existed among x column names.}

\item{scale}{type of data to plot: raw count, percentages or count per million.}
}
\value{
a heatmap
}
\description{
function plots a heatmap of V-J counts for sampleName
}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
