% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDissimilarityMatrix}
\alias{plotDissimilarityMatrix}
\title{create heatmap of pairwise distance matrix between samples}
\usage{
plotDissimilarityMatrix(
  x,
  level = c("VpJ", "V", "J", "VJ", "CDR3aa"),
  method = c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski",
    "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial",
    "chao", "cao", "mahalanobis"),
  binary = FALSE
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{level}{a string of characters indicating the level at which performed all calculations.}

\item{method}{distance computation method. Choose one among manhattan, euclidean, canberra, clark, bray, kulczynski, jaccard, gower, altGower, morisita, horn, mountford, raup, binomial, chao, cao, mahalanobis.}

\item{binary}{if TRUE data will be transformed into present/absent data.}
}
\value{
a heatmap
}
\description{
function plot a heatmap of a squared distance matrix computed between samples
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
# plot heatmap of dissimilarity matrix.
plotDissimilarityMatrix(x = RepSeqData, level = "V", method = "euclidean")
}
}
