% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDistribVpJ}
\alias{plotDistribVpJ}
\title{frequency of clonotype by its rank}
\usage{
plotDistribVpJ(x, colorBy = NULL, aggreg = c("sum", "mean"))
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{colorBy}{column name in sample Data indicating groups of samples}

\item{aggreg}{aggregation of counts by sum or by mean}
}
\value{
a plot
}
\description{
function
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
get the names of columns
names(sData(RepSeqData))
# plot mean distribution of clonotypes by project
plotDistribVpJ(x = RepSeqData, colorBy = "project", aggreg = "mean")
}
}
