% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotFreqVpJ}
\alias{plotFreqVpJ}
\title{plot frequency}
\usage{
plotFreqVpJ(x, sampleName = NULL)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{sampleName}{sample to be plotted. If NULL the first sample name is used.}
}
\value{
a bar plot
}
\description{
function plot the histogram of counts for a sample broken into 6 intervals 1, ]1, 10], ]10, 100], ]100, 1000], ]1000, 10000], ]10000, more].
}
