% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotMDS}
\alias{plotMDS}
\title{biplot Multidimensional Scaling result}
\usage{
plotMDS(
  x,
  level = c("VpJ", "V", "J", "VJ", "CDR3aa"),
  method = c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski",
    "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial",
    "chao", "cao", "mahalanobis"),
  colGrp = NULL
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment.}

\item{level}{repertoire level VpJ, V, J, VJ, CDR3aa.}

\item{method}{distance computation method. Choose one among manhattan, euclidean, canberra, clark, bray, kulczynski, jaccard, gower, altGower, morisita, horn, mountford, raup, binomial, chao, cao, mahalanobis.}

\item{colGrp}{color by group of samples. A string of character containng the name of the varibale indicating group of samples (must be a column name in the slot sampleData). If NULL, the number of colors is equal to the number of sample.}
}
\value{
a dot plot.
}
\description{
function computes pairwise distances between all libs for a chosen level. Then, multidimentional scaling (2 dimensions) is applied to the resulting matrix, ggplot was used to represente individuals on the scaling dimensions.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
colnames(sData(RepSeqData))
# plot 2 first axes of multidimensional scaling performed on the dissimilarity matrix.
plotMDS(x = RepSeqData, level = "V", method = "euclidean", colGrp = "project")
}
}
