% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotRenyiProfiles}
\alias{plotRenyiProfiles}
\title{plot Renyi's profiles}
\usage{
plotRenyiProfiles(
  x,
  alpha = c(0, 0.25, 0.5, 1, 2, 4, 8, 16, 32, 64, Inf),
  level = c("VpJ", "V", "J", "VJ", "CDR3aa"),
  colorBy = NULL
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{alpha}{a vector of parameters for estimating Renyi's entropy.}

\item{level}{level of repertoire to estimated}

\item{colorBy}{name of a factor in sample information data (run sData(x))}
}
\value{
a graph
}
\description{
function plots Renyi's entropy profile for all samples in a RepSeqExperiment object.
}
