% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotSpectratype}
\alias{plotSpectratype}
\title{plot spectratype}
\usage{
plotSpectratype(x, sampleName = NULL)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}

\item{sampleName}{sample to plot, if NULL the first sample of x is plotted.}
}
\value{
a barplot
}
\description{
function plots spectratype for V genes of a repertoire.
}
