% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotVenn}
\alias{plotVenn}
\title{plot Venn's diagram}
\usage{
plotVenn(x, level = c("V", "J", "VJ", "VpJ", "CDR3aa"), colorBy = NULL)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{level}{repertoire level}

\item{colorBy}{color by group of samples. If not NULL}
}
\value{
a Venn's diagram
}
\description{
function plot Venn's diagram up to 5 samples
}
