% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{rarefyRepSeqExp}
\alias{rarefyRepSeqExp}
\title{down-sampling to have the same library size.}
\usage{
rarefyRepSeqExp(
  x,
  sample.size = min(sData(x)$nReads),
  rngseed = FALSE,
  replace = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment.}

\item{sample.size}{an integer indicate the final library size for all samples. The default is the smallest library size among all samples of the data set.}

\item{rngseed}{a integer used as seed for a reproducible result.}

\item{replace}{a boolean if TRUE the resampling is drawn with replacement.}

\item{verbose}{a boolean  if TRUE the detail of computation step will be shown.}
}
\value{
an object of class RepSeqExperiment
}
\description{
function resamples counts to a lower same library size. Two resampling methods is proposed: with or without replacement.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
RepSeqDataDownSampling <- rarefyRepSeqExp(x = RepSeqData, rngseed = 1234)
RepSeqDataDownSampling
}
}
