% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{raretabRepSeq}
\alias{raretabRepSeq}
\title{compute rarefaction table for a RepSeqExperiment object}
\usage{
raretabRepSeq(x)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}
}
\value{
a data.table containing rarefaction table in order to be represented graphically as rarefaction curves.
}
\description{
function computes rarefaction table for each sample (libnames). The number of clontypes detected in the function of the number of sequences generated.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
raretab <- raretabRepSeq(x = RepSeqData)
raretab
}
}
