% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{readClonotypes}
\alias{readClonotypes}
\title{read clonotype table}
\usage{
readClonotypes(
  path,
  aligner = c("MiXCR", "Adaptive", "rTCR", "ClonotypeR"),
  chain = c("A", "B"),
  keep.ambiguous = FALSE,
  keep.unproductive = FALSE,
  aa.th = 8,
  outFiltered = F
)
}
\arguments{
\item{path}{a path to a clonotype file, for the moment, only TSV from ClonotypeR is supported.}

\item{aligner}{choice between type of input file provided by "Adaptive", "ClonotypeR", "MiXCR", "rTCR". Default is "MiXCR".}

\item{chain}{a character \code{A} or \code{B} indicating the chain to be imported. Default is \code{A}.}

\item{keep.ambiguous}{a boolean choice if ambiguous clonotypes (contain STOP codon) should be kept in analysis. Default is \code{FALSE}.}

\item{keep.unproductive}{a boolean choice if unproductive clonotypes (Euclidean dividion of aa length by 3 > 0) should be kept in analysis. Default is \code{FALSE}.}

\item{aa.th}{an interger indicates the maximal number of amino acids could be deviated from the mean length. Default is \code{8}.}

\item{outFiltered}{write dropped reads to file. Default folder is getwd().}
}
\value{
a filtered clonotype table.
}
\description{
function readColontypes is a wrapper of parseClonotypes & filterClonotypes
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
l <- list.files(system.file(file.path('extdata/mixcr'), 
                         package = 'RepSeqData'), 
                         full.names = TRUE)
l
# first element of the vector
path <- l[1]
dataset <- readClonotypes(path = path, 
                   aligner = "MiXCR", 
                   chain = "B", 
                   keep.ambiguous=FALSE, 
                   keep.unproductive=FALSE, 
                   aa.th=8, 
                   outFiltered=F)
dataset
}
}
\seealso{
\code{\link{parseClonotypeR}}, \code{\link{filterClonotypes}}
}
