% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{renyiProfiles}
\alias{renyiProfiles}
\title{compute Renyi's profile}
\usage{
renyiProfiles(
  x,
  scales = c(0, 0.25, 0.5, 1, 2, 4, 8, 16, 32, 64, Inf),
  level = c("VpJ", "V", "J", "VJ", "CDR3aa")
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{scales}{a numerical vector of alpha to evaluate Renyi's entropy}

\item{level}{level of repertoire to assess Renyi profiles}
}
\value{
a data.table containing Renyi's profiles computed for each repertoire
}
\description{
function computes Renyi profile for all samples in a RepSeqExperiment object
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
renyiNumbers <- renyiProfiles(RepSeqData, level = "VpJ")
renyiNumbers
}
}
