% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepSeq.R
\name{RepSeq}
\alias{RepSeq}
\title{RepSeq}
\description{
Tools for analyse of HTS RepSeq data.
}
\details{
This package is for analysing high-thoughput sequencing repertoire data. 
It focus on clonotype data sets pre-processd by ClonotypeR, rTCR and MiXCR. 
Clonotype tables from different samples are merged, filtered, normalized and 
tested for differential expression between groups of samples. 
Diversity indices are also studied.
Results could be visualized in different ways.
}
