% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{RepSeqExp}
\alias{RepSeqExp}
\title{build RepSeqExperiment object from clonotype tables}
\usage{
RepSeqExp(clonotypetab, sampleinfo = NULL)
}
\arguments{
\item{clonotypetab}{a data.table of clonotype tables. The column names should be: lib, V, J, CDR3aa, CDR3dna, VpJ, VJ, score, count.}

\item{sampleinfo}{a data frame containing sample meta data}
}
\value{
an object of class RepSeqExperiment
}
\description{
function creates RepSeqExperiment object from clonotype tables.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
l <- list.files(system.file(file.path('extdata/mixcr'), 
                     package = 'RepSeqData'), 
                     full.names = TRUE)
l# list of gz-compressed files 
sampleData <- read.table(system.file(file.path('extdata/sampledata.txt'), 
                     package='RepSeqData'), 
                     sep = "\t", 
                     row.names = 2) 
clontab.tmp <- lapply(l, readClonotypes, aligner="MiXCR")
clontab <- data.table::rbindlist(dataset.tmp)
dataset <- RepSeqExp(clontab, sampleinfo = sampleData)
dataset
}
}
