% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepSeqExperiment.R
\docType{class}
\name{RepSeqExperiment-class}
\alias{RepSeqExperiment-class}
\alias{RepSeqExperiment}
\alias{assay,RepSeqExperiment-method}
\alias{assay}
\alias{assay<-,RepSeqExperiment-method}
\alias{assay<-}
\alias{mData,RepSeqExperiment-method}
\alias{mData}
\alias{mData<-,RepSeqExperiment-method}
\alias{mData<-}
\alias{sData,RepSeqExperiment-method}
\alias{sData}
\alias{sData<-,RepSeqExperiment-method}
\alias{sData<-}
\alias{History,RepSeqExperiment-method}
\alias{History}
\alias{History<-,RepSeqExperiment-method}
\alias{History<-}
\alias{show,RepSeqExperiment-method}
\alias{names,RepSeqExperiment-method}
\alias{names}
\alias{names<-,RepSeqExperiment-method}
\alias{[,RepSeqExperiment,ANY,ANY,ANY-method}
\alias{[,RepSeqExperiment-method}
\alias{is.RepSeqExperiment}
\title{Class RepSeqExperiment}
\usage{
assay(object)

assay(object, i, j) <- value

mData(object)

mData(object) <- value

sData(object)

sData(object) <- value

History(object)

History(object) <- value

\S4method{assay}{RepSeqExperiment}(object)

\S4method{assay}{RepSeqExperiment}(object, i, j) <- value

\S4method{mData}{RepSeqExperiment}(object)

\S4method{mData}{RepSeqExperiment}(object) <- value

\S4method{sData}{RepSeqExperiment}(object)

\S4method{sData}{RepSeqExperiment}(object) <- value

\S4method{History}{RepSeqExperiment}(object)

\S4method{History}{RepSeqExperiment}(object) <- value

\S4method{show}{RepSeqExperiment}(object)

\S4method{names}{RepSeqExperiment}(x)

\S4method{names}{RepSeqExperiment}(x) <- value

\S4method{[}{RepSeqExperiment,ANY,ANY,ANY}(x, i, j, drop)

is.RepSeqExperiment(x)
}
\arguments{
\item{object}{a RepSeqExperiment object}

\item{i}{indice(s) of clonotype(s) to extract}

\item{j}{indice(s) of sample(s) to extract}

\item{value}{either numeric, character or data frame}

\item{x}{an object}

\item{drop}{If \code{TRUE} the result is coerced to the lowest possible dimension}
}
\value{
an object of class [\code{\linkS4class{RepSeqExperiment}}]
}
\description{
An S4 class to represent a Hight Throuput Sequencing Immune Repertoire Experiment.

\code{[} Extract parts of RepSeqExperiment.
}
\section{Slots}{

\describe{
\item{\code{assayData}}{a data.table containing information about clonotype.}

\item{\code{sampleData}}{a data frame containing sample information such as treatment groups, species, ...}

\item{\code{metaData}}{a list of meta data}

\item{\code{History}}{a data frame registering operations done on the object.}
}}

\author{
H. P. Pham (mailto: hp.pham(a)pareanbiotech.fr, #chokomlabreizh)
}
