% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{segmentUsage}
\alias{segmentUsage}
\title{compute usage of segments}
\usage{
segmentUsage(x, level = c("VpJ", "V", "J", "VJ"))
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}], returned by \code{\link{readClonotypeSet}}}

\item{level}{segment usage (ie percentage) of VpJ, V, J, or VJ within a repertoire.}
}
\value{
a data.table of segment usage of genes in each repertoire.
}
\description{
This function computed clonotype usage
}
