% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{toDESeq2}
\alias{toDESeq2}
\title{RepSeqCount to DESeqDataSet}
\usage{
toDESeq2(x, conditions, level = c("VpJ", "V", "J", "VJ", "CDR3aa"))
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{conditions}{name(s) of column(s) in sample information (sData(x)) defining test between biological conditions.}

\item{level}{of repertoire to analyze \code{VpJ, VJ, V, J, CDR3aa}.}
}
\value{
a DESeqDataSet object
}
\description{
function transforms a RepSeqCount object to an DESeqDataSet object
}
