% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{Chaowor}
\alias{Chaowor}
\title{Adjusted Chao1 for sampling without replacement}
\usage{
Chaowor(x)
}
\arguments{
\item{x}{a vector of counts}
}
\value{
a value of Chao1
}
\description{
function compute the correct
}
\examples{
set.seed(1234)
x <- rbinom(20, 5, 0.5)
y <- rbinom(20, 20, 0.1)
Chaowor(x)
Chaowor(y)
}
