% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{concateRepSeq}
\alias{concateRepSeq}
\title{concatenate 2 RepSeqExperiment objects}
\usage{
concateRepSeq(a, b)
}
\arguments{
\item{a}{first RepSeqExperiment object.}

\item{b}{second RepSeqExperiment object.}
}
\value{
a RepSeqExperiment object.
}
\description{
function concateRepSeq allow to concatenate 2 RepSeqExperiement objects. Overlaps in sample names are not allowed.
}
