% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filterFrequency}
\alias{filterFrequency}
\title{filter clonotypes presented in less than a proportion of samples.}
\usage{
filterFrequency(x, freq = 0.2, group = NULL)
}
\arguments{
\item{x}{an object of class RepSeqExperiment.}

\item{freq}{a number between 0 & 1. A clonotype is retained if it is present in a proportion of samples that excces This is the proportion of samples}

\item{group}{a factor in sampleData, if null, freq is applied to the group having the smallest number of membres.}
}
\value{
an object of class RepSeqExperiment satisfying the conditions.
}
\description{
A function that returns a function with values for A, p and na.rm bound to the specified values. The function takes a single vector, x, as an argument. When the returned function is evaluated it returns TRUE if the proportion of values in x that are larger than A is at least p.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeq)
library(RepSeqData)
sData(RepSeqData)
# get overlap clonotypes between the sample S01 and the sample S02
filterdata <- filterFrequency(RepSeqData, 
                       freq = 0.2, 
                       group = "quantity")  
filterdata
}
}
