% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{iChao}
\alias{iChao}
\title{Improved Chao1}
\usage{
iChao(x)
}
\arguments{
\item{x}{a vector of count.}
}
\value{
improved Chao1 value (ref) Chao and Lin Chao, A. and Lin, C.-W. (2012) Nonparametric lower bounds for species richness and shared species richness under sampling without replacement. Biometrics,68, 912–921.
}
\description{
function computes the improve version of Chao1
}
\examples{
set.seed(1234)
x <- rbinom(20, 5, 0.5)
y <- rbinom(20, 20, 0.1)
iChao(x)
iChao(y)
}
