% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{parseMiXCR}
\alias{parseMiXCR}
\title{parse MiXCR tables}
\usage{
parseMiXCR(path, chain = c("A", "B"))
}
\arguments{
\item{path}{full path to a file returned by MiXCR}

\item{chain}{TCR chain \code{A} or [code{B} to extract. Default value is \code{A}.}
}
\value{
a data.table
}
\description{
function imports clonotype tables produced by MiXCR aligner.
}
