% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{parseRTCR}
\alias{parseRTCR}
\title{parse rTCR tables}
\usage{
parseRTCR(path)
}
\arguments{
\item{path}{full path to a file returned by rTCR}
}
\value{
a data.table
}
\description{
function imports clonotype tables produced by rTCR aligner.
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
l <- list.files(system.file(file.path('extdata/rtcr'), 
                 package = 'RepSeqData'), 
                 full.names = TRUE)
l
# first element of the vector
path <- l[1]
dataset <- parseRTCR(path, chain = "B")
dataset
}
}
