% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot2v2count}
\alias{plot2v2count}
\title{sample against sample plot}
\usage{
plot2v2count(
  x,
  level = c("V", "J", "VJ", "VpJ", "CDR3aa"),
  libs = NULL,
  scale = c("counts", "log")
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{level}{choose between V, J, VJ, VpJ, CDR3aa}

\item{libs}{a vector of characters of length 2 containing the sample names to plot.}

\item{scale}{normal or log10 scale of axes}
}
\value{
a dot plot
}
\description{
function plots
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
rownames(sData(RepSeqData))
plot correlation of counts between 2 samples 
plot2v2count(x = RepSeqData, level = "V", libs = c("S01", "S02"), scale = "log")    
}
}
