% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotCountVJ}
\alias{plotCountVJ}
\title{plot a heatmap og V-J count}
\usage{
plotCountVJ(x, sampleName = NULL)
}
\arguments{
\item{x}{a matrix of count data}

\item{sampleName}{sample to plot, sample name must be existed among x column names.}
}
\value{
a heatmap of count
}
\description{
function plots a heatmap of V-J counts for sampleName
}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
