% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDistribVpJ}
\alias{plotDistribVpJ}
\title{function plotDistribVpJ}
\usage{
plotDistribVpJ(x, colorBy = NULL, aggreg = c("sum", "mean"))
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{colorBy}{column name in sample Data indicating groups of samples}

\item{aggreg}{aggregation of counts by sum or by mean}
}
\value{
a plot
}
\description{
function
}
