% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotFrequencySpectrum}
\alias{plotFrequencySpectrum}
\title{plot of frequency}
\usage{
plotFrequencySpectrum(x, colorBy = NULL, groupBy = FALSE)
}
\arguments{
\item{x}{x}

\item{colorBy}{color by factor name}

\item{groupBy}{boolean if compute the mean by factor or not}
}
\value{
a graph
}
\description{
function plots frequency of
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
# Show column names 
names(sData(RepSeqData))
plotFrequencySpectrum(x = RepSeqData, colorBy = "project", groupBy = TRUE)
}
}
