% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotmuScore}
\alias{plotmuScore}
\title{plot heatmap of counts or percentages}
\usage{
plotmuScore(x, level = c("V", "J", "VJ"), type = c("count", "usage"))
}
\arguments{
\item{x}{an object of class RepSeqExperiment}

\item{level}{of the repertoire}

\item{type}{count or usage}
}
\value{
a heatmap
}
\description{
function plots a heatmap of counts or percentages of sorted
}
