% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotPropVJ}
\alias{plotPropVJ}
\title{plot V or J proportion}
\usage{
plotPropVJ(
  x,
  level = c("V", "J"),
  sampleName = NULL,
  scale = c("counts", "percent", "cpm")
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment.}

\item{level}{level of repertoire to be used.}

\item{sampleName}{the name of sample or lib to be plotted, sample name must be existed among x column names. If NULL the first sample name is used.}

\item{scale}{type of data to be plotted, counts, percentages or counts per million.}
}
\value{
a barplot
}
\description{
function plots V or J proportion in all samples in a RepSeqExperiment object.
}
