% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotSpectratype}
\alias{plotSpectratype}
\title{plot spectratype}
\usage{
plotSpectratype(x, sampleName = NULL, scale = c("counts", "percent", "cpm"))
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}

\item{sampleName}{sample to plot, if NULL the first sample of x is plotted.}

\item{scale}{the type of bars in term of counts, percentages or count per million.}
}
\value{
a barplot
}
\description{
function plots spectratype for V genes of one repertoire (distribution of lengths of CDR3 in amio-acid).
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
library(RepSeqData)
snames <- rownames(sData(RepSeqData))
# plot Spectratype of the first sample, all V together.
plotSpectratype(x = RepSeqData, sampleName = snames[1], scale = "counts")
}
}
