% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotSpectratypeV}
\alias{plotSpectratypeV}
\title{plot Spectratype by V genes}
\usage{
plotSpectratypeV(
  x,
  sampleName = NULL,
  scale = c("counts", "percent", "cpm"),
  showCDR3 = F
)
}
\arguments{
\item{x}{an object of class RepSeqExperiment.}

\item{sampleName}{a string indicating the sample to plot. sampleName must be existing in the slot "sampleData".}

\item{scale}{the type of bars in term of counts, percentages or count per million.}

\item{showCDR3}{blabla}
}
\value{
a wrapped of barplots
}
\description{
function produces barplot according to the CDR3 length for each V gene.
}
