% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{readClonotypeSet}
\alias{readClonotypeSet}
\title{read clonotype table set}
\usage{
readClonotypeSet(fileList, cores = 1L, aligner = c("ClonotypeR",
  "rTCR", "MiXCR"), chain = c("A", "B"), sampleinfo = NULL,
  keep.ambiguous = FALSE, keep.unproductive = FALSE, aa.th = 8)
}
\arguments{
\item{fileList}{a list of file paths to import.}

\item{cores}{an interger number of cores to use. Default is 1.}

\item{aligner}{software used to align reads ("rTCR", "ClonotypeR" or "MiXCR").}

\item{chain}{an character indicates which TCR chain alpha or beta to import, Use \code{A} for alpha chain and \code{B} for beta chain. Default is \code{A}.}

\item{sampleinfo}{a data frame containing sample information for each clonotype file in fileList. The number of rows of sampleinfo must be identical to the number of file in fileList. A data frame containing If NULL}

\item{keep.ambiguous}{a boolean choice if ambiguous clonotypes (contain STOP codon) should be kept in analysis. Default is \code{FALSE}.}

\item{keep.unproductive}{a boolean choice if unproductive clonotypes (Euclidean dividion of aa length by 3 > 0) should be kept in analysis. Default is \code{FALSE}.}

\item{aa.th}{an interger indicates the maximal number of amino acids could be deviated from the mean length. Default is 8.}
}
\value{
an object of class RepSeqExperiment.
}
\description{
function reads all clonotype tables stored in a folder, filters reads and assay reads for each clonotype.
}
