% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{segmentUsage}
\alias{segmentUsage}
\title{compute usage of segments}
\usage{
segmentUsage(x, level = c("VpJ", "V", "J", "VJ"))
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}], returned by \code{\link{readClonotypeSet}}}

\item{level}{segment usage (ie percentage) of VpJ, V, J, or VJ within a repertoire.}
}
\value{
a data.table of segment usage of genes in each repertoire.
}
\description{
This function computed clonotype usage
}
\examples{
\dontrun{
# The package RepSeqData contains example datasets 
l <- list.files(system.file(file.path('extdata/mixcr'), 
                   package = 'RepSeqData'), 
                   full.names = TRUE)
l# list of gz-compressed files 
sampleData <- read.table(system.file(file.path('extdata/sampledata.txt'), 
                     package='RepSeqData'), 
                     sep = "\t", 
                     row.names = 2) 
dataset <- readClonotypeSet(fileList = l, 
                   cores=1L, 
                   aligner = "MiXCR", 
                   chain = "B", 
                   sampleinfo = sampleData, 
                   keep.ambiguous=FALSE, 
                   keep.unproductive=FALSE, 
                   aa.th=8)
dataset
# produce a count matrix of VJ
seg.use <- segmentUsage(x = dataset, level = "VJ")
seg.use
}
}
